add several pages for cart, contact, orderdetail, placeorder

+++ RealWorldApp/Models/Order.cs
@@ -10,12 +10,10 @@ namespace RealWorldApp.Models
         public string fullName { get; set; }
         public string address { get; set; }
         public string phone { get; set; }
-        public int orderTotal { get; set; }
+        public double orderTotal { get; set; }
         public DateTime orderPlaced { get; set; }
         public bool isOrderCompleted { get; set; }
         public int userId { get; set; }
         public List<OrderDetail> orderDetails { get; set; }
-
     }
-
 }
